+++ src/components/Camera.js
@@ -1,0 +1,30 @@
+import React from "react";
+import Webcam from "react-webcam";
+
+const videoConstraints = {
+  width: 1280,
+  height: 720,
+  facingMode: "user",
+};
+
+export default function Camera(props) {
+  return (
+    <Webcam
+      audio={false}
+      height={720}
+      screenshotFormat="image/jpeg"
+      width={720}
+      videoConstraints={videoConstraints}
+    >
+      {({ getScreenshot }) => (
+        <button
+          onClick={() => {
+            const imageSrc = getScreenshot();
+          }}
+        >
+          Capture photo
+        </button>
+      )}
+    </Webcam>
+  );
+}
